Fix bug with team multi select filter + Code clean up
- ISSUE: When applying a team filter before other filters,
the scatter plot would throw an NoneType error.
This was due to a None value being passed to the .isin method via
input_conf or input_div since these are the default values
if value property is not defined.
- FIX: Use input_team to filter the dataframe only which allowed the
removal of .isin with input_conf and input_div since these are already
chained to input_team via callbacks.